Update price~>decimal, edit_password

+++ db/migrate/20220907030541_create_cart_sessions.rb
@@ -2,7 +2,7 @@ class CreateCartSessions < ActiveRecord::Migration[7.0]
   def change
     create_table :cart_sessions do |t|
       t.references :user
-      t.integer :sum_money, default: 0
+      t.decimal :sum_money, default: 0
 
       t.timestamps
     end
