+++ db/migrate/20220913072151_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
       t.references :user
-      t.integer :sum_money
+      t.decimal :sum_money
       
       t.timestamps
     end
